Create Dockerfils in temp dir
Rather than having to copy the repo into a temp dir, we can create the
dockerfile (and the dockerignore) in a temp file and refer to those from
within the repo to copied into the container.

This also suggests using Bos for the cmd handling, to streamline error
handling by working within the result monad.