Add docs for build flags
I realized that by turning some of these flags into enums for the
supported values we get them pretty well self-documenting, and eliminate
most kinds of possible configuration error. It remains possible to
introduce configuration error by supplying an invalid sha, but this is
specialized enough that I think we can leave it is a followup.

In any case, the shas are probably only revelant to reproduce CI builds,
in which case we will automate production of the CLI command to run.